Image Blocks: Adding an image thumbnail selector
