+++ blocks/inspector-controls/select-control/index.js
@@ -14,16 +14,19 @@ import { withInstanceId } from '@wordpress/components';
 import BaseControl from './../base-control';
 import './style.scss';
 
-function SelectControl( { label, selected, help, instanceId, onBlur, options = [], ...props } ) {
+function SelectControl( { label, help, instanceId, onChange, options = [], ...props } ) {
 	const id = 'inspector-select-control-' + instanceId;
-	const onBlurValue = ( event ) => onBlur( event.target.value );
+	const onChangeValue = ( event ) => onChange( event.target.value );
 
+	// Disable reason: A select with an onchange throws a warning
+
+	/* eslint-disable jsx-a11y/no-onchange */
 	return ! isEmpty( options ) && (
 		<BaseControl label={ label } id={ id } help={ help }>
 			<select
 				id={ id }
 				className="blocks-select-control__input"
-				onBlur={ onBlurValue }
+				onChange={ onChangeValue }
 				aria-describedby={ !! help ? id + '__help' : undefined }
 				{ ...props }
 			>
@@ -31,7 +34,6 @@ function SelectControl( { label, selected, help, instanceId, onBlur, options = [
 					<option
 						key={ option.value }
 						value={ option.value }
-						selected={ option.value === selected }
 					>
 						{ option.label }
 					</option>
@@ -39,6 +41,7 @@ function SelectControl( { label, selected, help, instanceId, onBlur, options = [
 			</select>
 		</BaseControl>
 	);
+	/* eslint-enable jsx-a11y/no-onchange */
 }
 
 export default withInstanceId( SelectControl );
